+++ src/ChatClient.ts
@@ -43,11 +43,13 @@ import {
   MTrenewToken,
   MTupdatePushConfig,
 } from './__internal__/Consts';
+import { ErrorHandler } from './__internal__/ErrorHandler';
 import { Native } from './__internal__/Native';
 import { ChatContactManager } from './ChatContactManager';
 import {
   ChatConnectEventListener,
   ChatCustomEventListener,
+  ChatErrorEventListener,
   ChatMultiDeviceEventFromNumber,
   ChatMultiDeviceEventListener,
 } from './ChatEvents';
@@ -941,6 +943,32 @@ export class ChatClient extends BaseManager {
   public removeAllCustomListener(): void {
     chatlog.log(`${ChatClient.TAG}: removeAllCustomListener: `);
     this._customListeners.clear();
+  }
+
+  /**
+   * Add error listener.
+   *
+   * Monitor SDK internal errors.
+   */
+  public addErrorListener(listener: ChatErrorEventListener): void {
+    chatlog.log(`${ChatClient.TAG}: addErrorListener: `);
+    ErrorHandler.getInstance().listeners.add(listener);
+  }
+
+  /**
+   * Remove error listener.
+   */
+  public removeErrorListener(listener: ChatErrorEventListener): void {
+    chatlog.log(`${ChatClient.TAG}: removeErrorListener: `);
+    ErrorHandler.getInstance().listeners.delete(listener);
+  }
+
+  /**
+   * Remove all error listener.
+   */
+  public removeAllErrorListener(): void {
+    chatlog.log(`${ChatClient.TAG}: removeAllErrorListener: `);
+    ErrorHandler.getInstance().listeners.clear();
   }
 
   /**