+++ src/ChatContactManager.ts
@@ -20,6 +20,7 @@ import {
   MTremoveUserFromBlockList,
   MTsetContactRemark,
 } from './__internal__/Consts';
+import { ErrorHandler } from './__internal__/ErrorHandler';
 import { Native } from './__internal__/Native';
 import type { ChatContactEventListener } from './ChatEvents';
 import { chatlog } from './common/ChatConst';
@@ -76,9 +77,12 @@ export class ChatContactManager extends BaseManager {
           break;
 
         default:
-          throw new ChatError({
-            code: 1,
-            description: `This type is not supported. ` + contactEventType,
+          ErrorHandler.getInstance().sendError({
+            error: new ChatError({
+              code: 1,
+              description: `This type is not supported. ` + contactEventType,
+            }),
+            from: ChatContactManager.TAG,
           });
       }
     });