+++ src/ChatRoomManager.ts
@@ -34,6 +34,7 @@ import {
   MTunMuteChatRoomMembers,
   MTupdateChatRoomAnnouncement,
 } from './__internal__/Consts';
+import { ErrorHandler } from './__internal__/ErrorHandler';
 import { Native } from './__internal__/Native';
 import type { ChatRoomEventListener } from './ChatEvents';
 import { chatlog } from './common/ChatConst';
@@ -73,10 +74,6 @@ export class ChatRoomManager extends Native {
   private invokeRoomListener(params: any): void {
     this._roomListeners.forEach((listener: ChatRoomEventListener) => {
       const contactEventType = params?.type;
-      if (contactEventType === undefined) {
-        chatlog.warn('invokeRoomListener:invokeRoomListener:', params);
-        return;
-      }
       switch (contactEventType) {
         case 'onChatRoomDestroyed':
           listener.onDestroyed?.({
@@ -184,9 +181,12 @@ export class ChatRoomManager extends Native {
           break;
 
         default:
-          throw new ChatError({
-            code: 1,
-            description: `This type is not supported. ` + contactEventType,
+          ErrorHandler.getInstance().sendError({
+            error: new ChatError({
+              code: 1,
+              description: `This type is not supported. ` + contactEventType,
+            }),
+            from: ChatRoomManager.TAG,
           });
       }
     });